refactor: move clerk appearance variables to it's own hook
