+++ src/modules/auth/ui/components/clerk-user-button.tsx
@@ -1,16 +1,15 @@
 "use client";
 
 import { Skeleton } from "@/components/ui/skeleton";
-import { allThemes, type ThemeKey } from "@/config/themes";
+import { useClerkAppearanceVariables } from "@/modules/user/hooks/use-clerk-appearance-variables";
 import { UserButton, useAuth } from "@clerk/nextjs";
 import { Home, LayoutDashboard } from "lucide-react";
-import { useTheme } from "next-themes";
 import { useEffect, useState } from "react";
 
 export const ClerkUserButton = () => {
   const [isMounted, setIsMounted] = useState(false);
   const { isLoaded } = useAuth();
-  const { resolvedTheme } = useTheme();
+  const clerkAppearanceVariables = useClerkAppearanceVariables();
 
   useEffect(() => {
     setIsMounted(true);
@@ -19,26 +18,6 @@ export const ClerkUserButton = () => {
   if (!isMounted || !isLoaded) {
     return <Skeleton className="h-8 w-8 rounded-full" />;
   }
-
-  const themeKey = resolvedTheme;
-  const themeObject = allThemes[themeKey as ThemeKey] ?? allThemes.light;
-
-  const clerkAppearanceVariables = {
-    colorBackground: themeObject.card,
-    colorNeutral: themeObject.foreground,
-    colorText: themeObject.foreground,
-    colorDanger: themeObject.destructive,
-    colorPrimary: themeObject.primary,
-    borderRadius: themeObject.radius,
-    colorTextSecondary: themeObject.mutedForeground,
-    colorTextOnPrimaryBackground: themeObject.foreground,
-    colorInputBackground: themeObject.input,
-    colorInputText: themeObject.foreground,
-    colorSuccess: themeObject.primary,
-    // colorShimmer: themeObject.border,
-    colorWarning: themeObject.destructive,
-    fontFamily: themeObject.fontSans,
-  };
 
   if (parseFloat(clerkAppearanceVariables.borderRadius) >= parseFloat("1rem")) {
     return (
