+++ src/modules/auth/ui/views/auth-view.tsx
@@ -1,31 +1,10 @@
 "use client";
 
-import { allThemes, type ThemeKey } from "@/config/themes";
+import { useClerkAppearanceVariables } from "@/modules/user/hooks/use-clerk-appearance-variables";
 import { SignIn } from "@clerk/nextjs";
-import { useTheme } from "next-themes";
 
 export const AuthView = () => {
-  const { resolvedTheme } = useTheme();
-
-  const themeKey = resolvedTheme;
-  const themeObject = allThemes[themeKey as ThemeKey] ?? allThemes.light;
-
-  const clerkAppearanceVariables = {
-    colorBackground: themeObject.card,
-    colorNeutral: themeObject.foreground,
-    colorText: themeObject.foreground,
-    colorDanger: themeObject.destructive,
-    colorPrimary: themeObject.primary,
-    borderRadius: themeObject.radius,
-    colorTextSecondary: themeObject.mutedForeground,
-    colorTextOnPrimaryBackground: themeObject.foreground,
-    colorInputBackground: themeObject.input,
-    colorInputText: themeObject.foreground,
-    colorSuccess: themeObject.primary,
-    // colorShimmer: themeObject.border,
-    colorWarning: themeObject.destructive,
-    fontFamily: themeObject.fontSans,
-  };
+  const clerkAppearanceVariables = useClerkAppearanceVariables();
 
   return (
     <div className="mt-16 lg:mt-0">
